Feat/#7 detail

+++ src/main/java/com/finsight/backend/domain/vo/user/HoldingsVO.java
@@ -13,7 +13,7 @@ public class HoldingsVO {
     private String userId;
     private String productCode;
     private String productCategory;
-    private Long holdingsTotalPrice;
+    private Double holdingsTotalPrice;
     private Integer holdingsTotalQuantity;
     private String holdingsStatus; // holding, zero, terminated, expired
